[#41] test class for the Socrata filter process method.
Writing tests and having to switch from Unit to Kernel is time consuming.
